Index page now more of a main screen. Preset loading and display implementation.
-Further refined form to include an active element, for later sorting
between active and inactive buffs (and later adding those for character
display)
-Adjusted JavaScript functions to help create a display list on the
index page. The buff form is now a separate page of the display.
x- This has introduced a new problem of getting data to the buff page,
ran out of time to implement in this upload, still working on it though.
-Adjust Images slightly for change in functionality, made a load
presets button(and added JS function) to load the json data then
display it on the main page.
-Added search function, however, I'm not sure this is so useful for my
app. Will experiment with function.

-Added Json template from week 1 to gold folder, no work done on this
so far.