Project Reboot
Pausing work on advanced functionality to work on required functions in
a format for project 2.
Moved copies of work done for branch site-a to backup folder. Inserted
new index from JQM project, copied in new folders for css and JS. Added
old work to git-ignore list, but left it in the file structure for
reference locally.